add sast metadata service
